pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hexlet-code"
version = "0.1.0"
description = "Page analyzer"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "flask>=3.0.0",
  "gunicorn>=21.2.0",
  "python-dotenv>=1.0.0",
  "psycopg2-binary>=2.9.9",
  "validators>=0.24.0",
]

[dependency-groups]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.6.0",
]

[tool.setuptools]
packages = ["page_analyzer"]
